Extract editable install support
